mentor_adjustments
