+++ src/test/kotlin/podongdaeng/homecoming/CrawlingTests.kt
@@ -1,10 +1,11 @@
 package podongdaeng.homecoming
 
 import org.junit.jupiter.api.Test
+import podongdaeng.homecoming.clients.TerrorlessCrawlingService
 import podongdaeng.homecoming.controller.BasicController
 
 class CrawlingTests {
-    private val sel = BasicService.TerrorlessCrawlingService()
+    private val sel = TerrorlessCrawlingService()
     private val stationGpsController = BasicController()
     @Test
     fun test1() {
@@ -13,6 +14,6 @@ class CrawlingTests {
 
     @Test
     fun test2() {
-        stationGpsController.searchAddress(gpsLati = 37.3.toString(), gpsLong = 126.5.toString())
+        stationGpsController.parseBusInfo(gpsLati = 37.3.toString(), gpsLong = 126.5.toString())
     }
 }
